backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
transformers==4.41.2
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
pillow==10.3.0
pinecone == 7.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
